05_11_2025_MiniProject_ATS_System/requirements.txt
streamlit
python-dotenv

# LangChain + OpenRouter
langchain
langchain-openai
langgraph

# PDF Processing
PyPDF2

# Machine Learning / Text Similarity
scikit-learn
numpy

# Email
secure-smtplib